requirements.txt
biopython
numpy
pandas
tkinter